refactor\!: unify background task and theme systems
BREAKING CHANGE: Replace legacy implementations with optimized versions

- Theme System: Replace theme.py with optimized_theme.py (47% code reduction)
- Background Tasks: Replace background_tasks.py with simple_background_tasks.py (95% code reduction)
- Uses CSS variables for maintainable theming
- Uses Streamlit built-ins for simplified task management

This is a breaking change that will require import updates in dependent modules.
