--- src/ui/utils/simple_background_tasks.py
@@ -1,116 +1,0 @@
-"""Streamlined background task management using Streamlit built-ins.
-
-This module provides a library-first approach to background task management,
-replacing complex custom threading with st.status() + simple threading for
-optimal performance and maintainability.
-
-Key improvements:
-- 95% code reduction (806 → 50 lines)
-- Uses st.status() for better UX
-- Simple threading instead of ThreadPoolExecutor
-- Direct st.session_state integration
-- Enhanced database session management for background threads
-- No memory leaks or cleanup needed
-"""
-
-import logging
-import threading
-
-from typing import Any
-
-import streamlit as st
-
-from src.scraper import scrape_all
-from src.ui.utils.database_utils import (
-    clean_session_state,
-    suppress_sqlalchemy_warnings,
-)
-
-logger = logging.getLogger(__name__)
-
-# Suppress SQLAlchemy warnings common in Streamlit context
-suppress_sqlalchemy_warnings()
-
-
-def render_scraping_controls() -> None:
-    """Render scraping controls with progress tracking.
-
-    Uses library-first st.status() for progress visualization and
-    st.session_state for state management. Includes database session
-    cleanup to prevent contamination.
-    """
-    # Clean any contaminated database objects from session state
-    clean_session_state()
-
-    # Initialize scraping state
-    if "scraping_active" not in st.session_state:
-        st.session_state.scraping_active = False
-    if "scraping_results" not in st.session_state:
-        st.session_state.scraping_results = None
-
-    col1, col2 = st.columns([1, 1])
-
-    with col1:
-        if not st.session_state.scraping_active and st.button(
-            "🔍 Start Scraping", type="primary"
-        ):
-            start_scraping()
-
-    with col2:
-        if st.session_state.scraping_active and st.button(
-            "⏹️ Stop Scraping", type="secondary"
-        ):
-            st.session_state.scraping_active = False
-            st.rerun()
-
-
-def start_scraping() -> None:
-    """Start background scraping with Streamlit status tracking."""
-    st.session_state.scraping_active = True
-
-    # Create status container for progress tracking
-    status_container = st.empty()
-
-    def scraping_task():
-        try:
-            with (
-                status_container.container(),
-                st.status("🔍 Scraping job listings...", expanded=True) as status,
-            ):
-                # Update progress during scraping
-                st.write("📊 Initializing scraping workflow...")
-
-                # Execute scraping (preserves existing scraper.py logic)
-                result = scrape_all()
-
-                # Show completion
-                total_jobs = sum(result.values()) if result else 0
-                status.update(
-                    label=f"✅ Scraping Complete! Found {total_jobs} jobs",
-                    state="complete",
-                )
-
-                # Store results
-                st.session_state.scraping_results = result
-                st.session_state.scraping_active = False
-
-        except Exception:
-            with status_container.container():
-                st.error("❌ Scraping failed")
-            st.session_state.scraping_active = False
-            logger.exception("Scraping failed")
-
-    # Start background thread (preserves non-blocking behavior)
-    thread = threading.Thread(target=scraping_task, daemon=True)
-    thread.start()
-
-
-# Simple API functions (preserve compatibility)
-def is_scraping_active() -> bool:
-    """Check if scraping is currently active."""
-    return st.session_state.get("scraping_active", False)
-
-
-def get_scraping_results() -> dict[str, Any]:
-    """Get results from the last scraping operation."""
-    return st.session_state.get("scraping_results", {})
